requirements.txt
# MoveMate backend — Python 3.12+
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
bcrypt>=4.2.0
PyJWT>=2.9.0
cryptography>=43.0.0
python-multipart>=0.0.12
httpx>=0.27.0
